테스트 분리 - MSA
